Moved student/training/exercise/ delivery to a new folder (delivery/)
