+++ src/pages/student/routes.tsx
@@ -4,7 +4,7 @@ import { TrainingTOCPageContainer } from './training/toc/pageContainer';
 import { TrainingListPage } from './training/list/page';
 import { studentRouteEnums } from '../../common/routeEnums/student';
 import { ExerciseAssessmentPage } from './training/exercise/assessment/page';
-import { ExerciseDeliveryPage } from './training/exercise/page';
+import { ExerciseDeliveryPage } from './training/exercise/delivery/page';
 
 // http://stackoverflow.com/questions/35048738/react-router-import-routes
 // AssembliesRoutes.js
@@ -19,9 +19,11 @@ export const StudentsRoutes = (
     <Redirect from={studentRouteEnums.training.id.base} to={studentRouteEnums.training.id.toc} />
 
     {/* Mocked links */}
-    <Route path={`${studentRouteEnums.training.base}/1/exercise`} component={ExerciseAssessmentPage} />
     <Route path="/student/training/1/exercise/1/delivery" component={ExerciseDeliveryPage} />
     <Route path="/student/training/1/exercise/2/delivery" component={ExerciseDeliveryPage} />
     <Route path="/student/training/1/exercise/3/delivery" component={ExerciseDeliveryPage} />
+    <Route path={`${studentRouteEnums.training.base}/1/exercise`} component={ExerciseAssessmentPage} />
+    <Route path={`${studentRouteEnums.training.base}/2/exercise`} component={ExerciseAssessmentPage} />
+    <Route path={`${studentRouteEnums.training.base}/3/exercise`} component={ExerciseAssessmentPage} />
   </div>
 );
