+++ src/pages/student/training/exercise/delivery/components/droppableFile.scss
